Changed placements of specific Things and Connections implementations - platform-specific implementations are now moved to specific/platforms folder - each implementation is placed to corresponding [platform_name] folder - each [platform_name] folder contains an implementation of corresponding   connection and thing classes
