+++ dpl/specific/platforms/mpd/__init__.py
@@ -1,0 +1,4 @@
+from .mpd_client import MPDClientConnection
+from .mpd_player import MPDPlayer
+
+__all__ = ["MPDClientConnection", "MPDPlayer"]
