+++ dpl/specific/platforms/shift_reg_gpio/__init__.py
@@ -1,0 +1,5 @@
+from .shift_reg_gpio_buffered import ShiftRegGPIOBuffered
+from .shift_reg_slider import ShiftRegSlider
+from .shift_reg_trigger import ShiftRegTrigger
+
+__all__ = ["ShiftRegGPIOBuffered", "ShiftRegSlider", "ShiftRegTrigger"]
