+++ dpl/specific/platforms/shift_reg_gpio/shift_reg_gpio_buffered.py
