--- dpl/specific/things/__init__.py
@@ -1,5 +1,0 @@
-from .players import MPDPlayer
-from .sliders import ShiftRegSlider
-from .triggers import ShiftRegTrigger
-
-__all__ = ["MPDPlayer", "ShiftRegSlider", "ShiftRegTrigger"]
