--- dpl/specific/things/players/__init__.py
@@ -1,1 +1,0 @@
-from .mpd_player import MPDClientConnection, MPDPlayer
