--- dpl/specific/things/sliders/__init__.py
@@ -1,1 +1,0 @@
-from .shift_reg_slider import ShiftRegSlider
