--- dpl/specific/things/triggers/__init__.py
@@ -1,1 +1,0 @@
-from .shift_reg_trigger import ShiftRegTrigger
